projects/services/bg-ingest/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bg-ingest"
version = "0.1.0"
description = "Service for ingesting blood glucose readings from CGM providers"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=2.0.0",
    "boto3>=1.24.0",
    "httpx>=0.23.0",
    "prometheus_client",
    "pydantic-settings",
    "PyJWT>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=2.12.0",
    "moto>=3.1.0",
    "black>=22.1.0",
    "isort>=5.10.0",
    "mypy>=0.961",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100 
